\name{anRpackage-package}
\alias{anRpackage-package}
\alias{BurgerBurnin}
\docType{package}
\title{
This application will adjust your car's miliage based on your beef consumption.
~~ BurgerBurnin ~~
}
\description{
It demonstrates that the amount of beef that you consume affects the ammount of gasoline that you use almost as much as what kind of car you drive.  It uses the widely accepted fact that it takes one gallon of gasoline to produce one pound of grain fed beef.
}
\details{
\tabular{ll}{
Package: \tab BurgerBurnin\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-08-24\cr
License: \tab GNU\cr
}
This is a simple web app that illustrates alternative ways in which you can calculate how much gasoline that one is using when dining out in a restaurant.  Since it raising beef consumes a lot of gasoline, we adjust the miliage of your car by adding in the gasoline used to produce any beef related food that is consumed.  Cheers!
}
\author{
wporterjr

Maintainer: woz300z@yahoo.com
}
\references{
http://en.wikipedia.org/wiki/Environmental_impact_of_meat_production
https://alumni.stanford.edu/get/page/magazine/article/?article_id=29892
http://ajcn.nutrition.org/content/78/3/660S.full
http://www.globalissues.org/article/240/beef
http://www.worldwatch.org/node/549
}
\keyword{ package }

\seealso{
Google
}

